tests/wpt/mozilla/tests/mozilla/layout_blocker_operations.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="foo"></div>

<script>
  function verifyLayoutAllowed(t) {
      t.step(() => {
          assert_equals(getComputedStyle(document.querySelector('#foo')).display, "block");
          t.done();
      })
  }

  var insertionTest;
  async_test(function(t) {
      insertionTest = t;
      let script = document.createElement('script');
      document.body.appendChild(script);
      script.appendChild(document.createTextNode("verifyLayoutAllowed(insertionTest)"));
  }, "Insertion");

  var replacementTest;
  async_test(function(t) {
      replacementTest = t;
      let script = document.createElement('script');
      document.body.appendChild(script);
      script.innerHTML = "verifyLayoutAllowed(replacementTest)";
  }, "Replace");
</script>
</body>
</html>
